requirements.txt
torch
torchvision
fastapi
uvicorn
opencv-python-headless
pillow
numpy
python-multipart